restructured polkadb package and added more opts to badger
