+++ polkadb/database_test.go
@@ -18,8 +18,6 @@ package polkadb
 
 import (
 	"bytes"
-	"io/ioutil"
-	"os"
 	"testing"
 
 	"fmt"
@@ -30,43 +28,6 @@ import (
 type data struct {
 	input    string
 	expected string
-}
-
-func newTestDBService() (*DbService, func()) {
-	dir, err := ioutil.TempDir(os.TempDir(), "test_data")
-	if err != nil {
-		panic("failed to create test file: " + err.Error())
-	}
-	db, err := NewDatabaseService(dir)
-	if err != nil {
-		panic("failed to create test database: " + err.Error())
-	}
-	return db, func() {
-		db.Stop()
-		if err := os.RemoveAll(dir); err != nil {
-			fmt.Println("removal of temp directory test_data failed")
-		}
-	}
-}
-
-func TestDbService_Start(t *testing.T) {
-	db, remove := newTestDBService()
-	defer remove()
-
-	err := db.Start()
-	if err == nil {
-		t.Fatalf("get returned wrong result, got %v", err)
-	}
-}
-
-func TestDb_Close(t *testing.T) {
-	db, remove := newTestDBService()
-	defer remove()
-
-	success := db.StateDB.Db.Close()
-	if !success {
-		t.Fatalf("get returned wrong result, got %v expected %v", success, true)
-	}
 }
 
 func testSetup() []data {
@@ -156,7 +117,7 @@ func testDelGetter(db Database, t *testing.T) {
 			if err != nil {
 				t.Fatalf("delete %q failed: %v", v.input, err)
 			}
-			d, err := db.Get([]byte(v.input))
+			d, _ := db.Get([]byte(v.input))
 			if len(d) > 1 {
 				t.Fatalf("failed to delete value %q", v.input)
 			}
@@ -164,7 +125,7 @@ func testDelGetter(db Database, t *testing.T) {
 	}
 }
 
-func testGetPath(db *Db, t *testing.T) {
+func testGetPath(db Database, t *testing.T) {
 	dir := db.Path()
 	if len(dir) <= 0 {
 		t.Fatalf("failed to set database path")
@@ -177,7 +138,7 @@ func TestBadgerDB_Batch(t *testing.T) {
 	testBatchPut(db.StateDB.Db, t)
 }
 
-func batchTestSetup(db *Db) (func(i int) []byte, func(i int) []byte, Batch) {
+func batchTestSetup(db Database) (func(i int) []byte, func(i int) []byte, Batch) {
 	testKey := func(i int) []byte {
 		return []byte(fmt.Sprintf("%04d", i))
 	}
@@ -188,7 +149,7 @@ func batchTestSetup(db *Db) (func(i int) []byte, func(i int) []byte, Batch) {
 	return testKey, testValue, b
 }
 
-func testBatchPut(db *Db, t *testing.T) {
+func testBatchPut(db Database, t *testing.T) {
 	k, v, b := batchTestSetup(db)
 
 	for i := 0; i < 10000; i++ {
@@ -224,7 +185,7 @@ func TestBadgerDB_Iterator(t *testing.T) {
 	testSeekKeyValueIterator(db.StateDB.Db, t)
 }
 
-func testIteratorSetup(db *Db, t *testing.T) {
+func testIteratorSetup(db Database, t *testing.T) {
 	k, v, b := batchTestSetup(db)
 
 	for i := 0; i < 5; i++ {
@@ -239,7 +200,7 @@ func testIteratorSetup(db *Db, t *testing.T) {
 	}
 }
 
-func testNewIterator(db *Db, t *testing.T) {
+func testNewIterator(db Database, t *testing.T) {
 	testIteratorSetup(db, t)
 
 	it := db.NewIterator()
@@ -264,7 +225,7 @@ func testNewIterator(db *Db, t *testing.T) {
 	}
 }
 
-func testNextKeyIterator(db *Db, t *testing.T) {
+func testNextKeyIterator(db Database, t *testing.T) {
 	testIteratorSetup(db, t)
 
 	it := db.NewIterator()
@@ -296,7 +257,7 @@ func testKVData() []data {
 	return testKeyValue
 }
 
-func testSeekKeyValueIterator(db *Db, t *testing.T) {
+func testSeekKeyValueIterator(db Database, t *testing.T) {
 	testIteratorSetup(db, t)
 	kv := testKVData()
 
@@ -319,135 +280,5 @@ func testSeekKeyValueIterator(db *Db, t *testing.T) {
 				t.Fatalf("failed to retrieve presented key, got %v, expected %v", it.Key(), k.expected)
 			}
 		})
-	}
-}
-
-func TestBadgerDB_TablePrefixOps(t *testing.T) {
-	db, remove := newTestDBService()
-	defer remove()
-
-	testPutTablesWithPrefix(db.StateDB.Db, t)
-	testHasTablesWithPrefix(db.StateDB.Db, t)
-	testDelTablesWithPrefix(db.StateDB.Db, t)
-	testTableClose(db.StateDB.Db, t)
-	testNewTableBatch(db.StateDB.Db, t)
-}
-
-func testPutTablesWithPrefix(db Database, t *testing.T) {
-	data := testKVData()
-	ops := NewTable(db, "99")
-
-	for _, v := range data {
-		v := v
-		t.Run("PutTablesWithPrefix", func(t *testing.T) {
-			err := ops.Put([]byte(v.input), []byte(v.expected))
-			if err != nil {
-				t.Fatalf("put failed: %v", err)
-			}
-			data, err := ops.Get([]byte(v.input))
-			if err != nil {
-				t.Fatalf("get failed: %v", err)
-			}
-			if !bytes.Equal(data, []byte(v.expected)) {
-				t.Fatalf("get returned wrong result, got %q expected %q", string(data), v.expected)
-			}
-		})
-	}
-}
-
-func testHasTablesWithPrefix(db Database, t *testing.T) {
-	data := testKVData()
-	ops := NewTable(db, "99")
-
-	for _, v := range data {
-		exists, err := ops.Has([]byte(v.input))
-		if err != nil {
-			t.Fatalf("has operation failed: %v", err)
-		}
-		if !exists {
-			t.Fatalf("has operation returned wrong result, got %t expected %t", exists, true)
-		}
-	}
-}
-
-func testDelTablesWithPrefix(db Database, t *testing.T) {
-	data := testKVData()
-	ops := NewTable(db, "99")
-
-	for _, v := range data {
-		v := v
-		t.Run("PutTablesWithPrefix", func(t *testing.T) {
-			err := ops.Del([]byte(v.input))
-			if err != nil {
-				t.Fatalf("delete %q failed: %v", v.input, err)
-			}
-			d, err := ops.Get([]byte(v.input))
-			if len(d) > 1 {
-				t.Fatalf("failed to delete value %q", v.input)
-			}
-		})
-	}
-}
-
-func testTableClose(db Database, t *testing.T) {
-	ops := NewTable(db, "99")
-
-	success := ops.Close()
-	if !success {
-		t.Fatalf("get returned wrong result, got %v expected %v", success, true)
-	}
-}
-
-func testNewTableBatch(db Database, t *testing.T) {
-	ops := NewTable(db, "99")
-	b := ops.NewBatch()
-
-	_, ok := b.(Batch)
-	if !ok {
-		t.Fatalf("get returned wrong result, got %v", ok)
-	}
-}
-
-func TestBadgerDB_TableBatchWithPrefix(t *testing.T) {
-	db, remove := newTestDBService()
-	defer remove()
-	testBatchTablePutWithPrefix(db.StateDB.Db, t)
-}
-
-func batchTableWithPrefixTestSetup(db *Db) (func(i int) []byte, func(i int) []byte, Batch) {
-	testKey := func(i int) []byte {
-		return []byte(fmt.Sprintf("%04d", i))
-	}
-	testValue := func(i int) []byte {
-		return []byte(fmt.Sprintf("%05d", i))
-	}
-	b := NewTableBatch(db, "98")
-	return testKey, testValue, b
-}
-
-func testBatchTablePutWithPrefix(db *Db, t *testing.T) {
-	k, v, b := batchTableWithPrefixTestSetup(db)
-
-	for i := 0; i < 10000; i++ {
-		err := b.Put(k(i), v(i))
-		if err != nil {
-			t.Fatalf("failed to add key-value to batch mapping  %q", err)
-		}
-		err = b.Write()
-		if err != nil {
-			t.Fatalf("failed to write batch %q", err)
-		}
-		size := b.ValueSize()
-		if size == 0 {
-			t.Fatalf("failed to set size of data in each batch, got %v", size)
-		}
-		err = b.Delete([]byte(k(i)))
-		if err != nil {
-			t.Fatalf("failed to delete batch key %v", k(i))
-		}
-		b.Reset()
-		if b.ValueSize() != 0 {
-			t.Fatalf("failed to reset batch mapping to zero, got %v, expected %v", b.ValueSize(), 0)
-		}
 	}
 }
